[10.2] - added get place functionality

+++ lib/domain/repository/place_repository.dart
@@ -2,7 +2,7 @@ import 'package:places/data/model/place.dart';
 
 abstract class PlaceRepository {
   Future<Place> createPlace();
-  Future<List<Place>> getPlaces(int radius, String category);
+  Future<List<Place>> getPlaces(double radius, String category);
   Future<Place> getPlaceById(int id);
   Future updatePlace(int id);
   Future deletePlace(int id);
